Fixup icons
